refactor: rename CoinMarketCap to cmc
