+++ packages/web3-providers/src/cmc/constants.ts
