+++ packages/web3-providers/src/cmc/mirrored_tokens.json
