+++ packages/web3-providers/src/index.ts
@@ -21,7 +21,7 @@ import { AlchemyEVM_API, AlchemyFlowAPI } from './alchemy'
 import { EthereumWeb3API } from './web3'
 import { MagicEdenAPI } from './MagicEden'
 import { TokenViewAPI } from './token-view'
-import { CoinMarketCapAPI } from './CoinMarketCap'
+import { CoinMarketCapAPI } from './cmc'
 import { UniSwapAPI } from './uniswap'
 import { RabbyAPI } from './rabby'
 import { GemAPI } from './gem'
